Cleanup code with scalac flags and WartRemover (#88)
* Add additional compiler flags to tighten up code
* Add WartRemover
* remove possibly confusing dot at the end of comment
* Bump Travis 2.12.x version
* Bump Scalatest version

+++ enumeratum-play-json/src/test/scala/enumeratum/values/EnumFormatsSpec.scala
@@ -21,7 +21,7 @@ class EnumFormatsSpec extends FunSpec with Matchers with EnumJsonFormatHelpers {
       JsString(s"$c")
     })
     testReads("ByteEnum", Bites, { b: Byte =>
-      JsNumber(b)
+      JsNumber(b.toInt)
     })
 
   }
@@ -36,7 +36,7 @@ class EnumFormatsSpec extends FunSpec with Matchers with EnumJsonFormatHelpers {
       JsString(s"$c")
     })
     testWrites("ByteEnum", Bites, { b: Byte =>
-      JsNumber(b)
+      JsNumber(b.toInt)
     })
 
   }
